LayoutTests/fast/forms/select-max-length-expected.txt

CONSOLE MESSAGE: Unable to expand the option list to length 100001 items. The maximum number of items allowed is 100000.
CONSOLE MESSAGE: Unable to expand the option list and set an option at index. The maximum list length is 100000.
CONSOLE MESSAGE: Unable to expand the option list and set an option at index. The maximum list length is 100000.
This test that setting HTMLSelectElement.length is capped to 100,000, and you can't add additional Option elements too.

PASS sel.length is 0
Trying: - sel.length = 100001;
PASS sel.length is 0
Trying: - sel.add(new Option, 0);
PASS sel.length is 1
Trying: - sel.length = 0;
PASS sel.length is 0
Index setter:
PASS sel[100001] = new Option(); sel.length is 0
PASS sel.options[100001] = new Option(); sel.length is 0
PASS successfullyParsed is true

TEST COMPLETE
